Mostly all here
Added archival system, and proper database entry. Will still need to determine a way to have it run according to my new outline.

Just need to figure out how to have it so once a DOI has been in database for 3 weeks with no abstract/title added, try to obtain title/abstract from RSS feed.

The problem here is that how to I know which archive to look in which contains the DOI. Have a timestamp, but would need to look through all archives for that day.