Separated RSS and PubMed Query
Have code to deposit DOI with archive filename completed for the most part. Couldn't get requests to return anything due to Unicode errors, not sure if it was just when saving to the file.(Works in Repl, but not PyCharm can but not sure how to incorporate getting raw xml from passing url to requests instead of feedparserDict(commented out and can change when I get a solution))

Have rough outline for logic of code for querying the database for doi's and converting doi to pmid and querying PubMed for titles/abstract. Also, when doi's have been in db for >3 weeks how to get file to try to parse the RSS feeds for titles/abstracts.