Proper Raw XML File Output Working
Next steps are that just need to:
- Archive title/abstract parsing
- Elsevier DOI's from CrossRef using titles
- Returning stats of script run at the end (ex. how many: added DOI's to database?, updated DOI's in DB w/ title/abstract?)

Prior to the next steps, all the RSS feeds from NP RSS Feeds spreadsheet need to be added to the rss_feed.txt file (contains each URL link on a separate line). This will allow for better logic for parsing out the archived titles and abstracts. Also, the URL link to Elsevier journals would be included in newly added feed, which is required for figuring out how to gathering DOIs from CrossRef using titles parsed from the RSS feeds (Ideally working simultaneously when parsing RSS feeds for DOI's to add the list of DOIs for each journal, but may need separate functions).